Add very experimental ARM NEON support, with sse2neon macro library: * support for _mm_minpos_epu16 missing still, so no align module * prefilter is computing different results * we probably have to check every single char, since x86 char != arm char * _mm_extract_epi8 _mm_srli_epi64 _mm_slli_epi64 _mm_mulhi_epu16 _mm_mul_epu32 _mm_cmpeq_epi32   Are untested just implemented analogous to already existing functions * no idea what else might be broken

+++ src/alignment/StripedSmithWaterman.cpp
@@ -707,7 +707,7 @@ void SmithWaterman::ssw_init (const Sequence* q,
 			compositionBias = (static_cast<int8_t>(compositionBias) < profile->composition_bias[i])
 							  ? compositionBias  :  profile->composition_bias[i];
 		}
-		compositionBias = std::min(compositionBias, 0);
+		compositionBias = std::min(compositionBias, static_cast<int32_t>(0));
 //		std::cout << compositionBias << std::endl;
 	}else{
 		memset(profile->composition_bias, 0, q->L* sizeof(int8_t));
